+++ src/test/kotlin/com/cmgapps/gradle/NpmVersionTaskShould.kt
@@ -6,17 +6,45 @@
 
 package com.cmgapps.gradle
 
+import com.cmgapps.gradle.model.NpmResponse
+import com.cmgapps.gradle.reporter.asString
+import com.cmgapps.gradle.service.NetworkService
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.mock.MockEngine
+import io.ktor.client.engine.mock.respond
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpStatusCode
+import io.ktor.http.headersOf
+import io.ktor.utils.io.ByteReadChannel
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
+import org.gradle.api.Action
 import org.gradle.api.Project
+import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.provider.MapProperty
+import org.gradle.api.provider.Property
+import org.gradle.kotlin.dsl.mapProperty
+import org.gradle.kotlin.dsl.property
 import org.gradle.testfixtures.ProjectBuilder
+import org.gradle.workers.ClassLoaderWorkerSpec
+import org.gradle.workers.ProcessWorkerSpec
+import org.gradle.workers.WorkAction
+import org.gradle.workers.WorkParameters
+import org.gradle.workers.WorkQueue
+import org.gradle.workers.WorkerExecutor
+import org.gradle.workers.WorkerSpec
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.`is`
-import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinMetadataTarget
 import org.jetbrains.kotlin.gradle.targets.js.npm.NpmDependency
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
+import java.io.ByteArrayOutputStream
+import java.io.PrintStream
 import java.nio.file.Path
+
+private const val TEST_DEP_NAME = "npm-dependency"
+private const val TEST_DEP_VERSION = "1.0.0"
 
 class NpmVersionTaskShould {
     @TempDir
@@ -32,30 +60,88 @@ class NpmVersionTaskShould {
                 .withProjectDir(testProjectDir.toFile())
                 .build()
 
-        project.plugins.apply(KotlinMetadataTarget.METADATA_TARGET_NAME)
+        project.plugins.apply("org.jetbrains.kotlin.multiplatform")
     }
 
     @Test
-    @Disabled("KotlinMultiplatformPlugin cannot be applied")
-    fun `setup correctly`() {
+    fun `create console report`() {
         val outputDir = testProjectDir.resolve("output")
         val task =
             project.tasks.create(
                 "npmVersions",
                 NpmVersionTask::class.java,
+                TestWorkExecutor(project),
+                project.objects,
             )
 
         val configuration = project.configurations.create("implementation")
 
-        configuration.dependencies.add(
+        project.dependencies.add(
+            "implementation",
             NpmDependency(
                 objectFactory = project.objects,
-                name = "npm-dependency",
-                version = "1.0.0",
+                name = TEST_DEP_NAME,
+                version = TEST_DEP_VERSION,
             ),
         )
 
-        task.outputDirectory.set(outputDir.toFile())
+        task.outputDirectory.set(outputDir.toFile().apply { mkdirs() })
+        task.configurationToCheck(
+            project.provider {
+                configuration
+            },
+        )
+
+        val outputStream = ByteArrayOutputStream()
+
+        System.setOut(PrintStream(outputStream))
+
+        task.action()
+
+        assertThat(
+            outputStream.asString(),
+            `is`(
+                "┌──────────────┐\n" +
+                    "│ NPM Packages │\n" +
+                    "└──────────────┘\n" +
+                    "\n" +
+                    "The following packages are using the latest version:\n" +
+                    " • npm-dependency:1.0.0\n",
+            ),
+        )
+    }
+
+    @Test
+    fun `create plain text report`() {
+        val outputDir = testProjectDir.resolve("output")
+        val task =
+            project.tasks.create(
+                "npmVersions",
+                NpmVersionTask::class.java,
+                TestWorkExecutor(project),
+                project.objects,
+            )
+
+        val configuration = project.configurations.create("implementation")
+
+        project.dependencies.add(
+            "implementation",
+            NpmDependency(
+                objectFactory = project.objects,
+                name = TEST_DEP_NAME,
+                version = TEST_DEP_VERSION,
+            ),
+        )
+
+        task.outputDirectory.set(outputDir.toFile().apply { mkdirs() })
+
+        val reportOutputFile = outputDir.resolve("report.txt")
+
+        task.reports.plainText.required
+            .set(true)
+        task.reports.plainText.outputLocation
+            .set(reportOutputFile.toFile())
+
         task.configurationToCheck(
             project.provider {
                 configuration
@@ -64,6 +150,177 @@ class NpmVersionTaskShould {
 
         task.action()
 
-        assertThat(outputDir.toFile().exists(), `is`(true))
+        assertThat(
+            reportOutputFile.toFile().readText(),
+            `is`(
+                "┌──────────────┐\n" +
+                    "│ NPM Packages │\n" +
+                    "└──────────────┘\n" +
+                    "\n" +
+                    "The following packages are using the latest version:\n" +
+                    " • npm-dependency:1.0.0\n",
+            ),
+        )
     }
+
+    @Test
+    fun `create json report`() {
+        val outputDir = testProjectDir.resolve("output")
+        val task =
+            project.tasks.create(
+                "npmVersions",
+                NpmVersionTask::class.java,
+                TestWorkExecutor(project),
+                project.objects,
+            )
+
+        val configuration = project.configurations.create("implementation")
+
+        project.dependencies.add(
+            "implementation",
+            NpmDependency(
+                objectFactory = project.objects,
+                name = TEST_DEP_NAME,
+                version = TEST_DEP_VERSION,
+            ),
+        )
+
+        task.outputDirectory.set(outputDir.toFile().apply { mkdirs() })
+
+        val reportOutputFile = outputDir.resolve("report.json")
+
+        task.reports.json.required
+            .set(true)
+        task.reports.json.outputLocation
+            .set(reportOutputFile.toFile())
+
+        task.configurationToCheck(
+            project.provider {
+                configuration
+            },
+        )
+
+        task.action()
+
+        assertThat(
+            reportOutputFile.toFile().readText(),
+            `is`(
+                "{\n" +
+                    "    \"latest\": [\n" +
+                    "        {\n" +
+                    "            \"name\": \"npm-dependency\",\n" +
+                    "            \"version\": \"1.0.0\"\n" +
+                    "        }\n" +
+                    "    ],\n" +
+                    "    \"outdated\": []\n" +
+                    "}\n",
+            ),
+        )
+    }
+}
+
+private class TestWorkExecutor(
+    project: Project,
+) : WorkerExecutor {
+    val workQueue =
+        object : WorkQueue {
+            override fun <T : WorkParameters?> submit(
+                workActionClass: Class<out WorkAction<T>>?,
+                parameterAction: Action<in T>?,
+            ) {
+                assertThat(workActionClass, `is`(CheckNpmPackageAction::class.java))
+
+                val action = TestCheckNpmPackageAction(project)
+
+                @Suppress("UNCHECKED_CAST")
+                parameterAction?.execute(action.parameters as T)
+                action.parameters.networkService.set(TestNetworkService(project))
+
+                action.execute()
+            }
+
+            override fun await() {
+            }
+        }
+
+    override fun noIsolation(): WorkQueue = workQueue
+
+    override fun noIsolation(action: Action<in WorkerSpec>?): WorkQueue = workQueue
+
+    override fun classLoaderIsolation(): WorkQueue {
+        TODO("Not yet implemented")
+    }
+
+    override fun classLoaderIsolation(action: Action<in ClassLoaderWorkerSpec>?): WorkQueue {
+        TODO("Not yet implemented")
+    }
+
+    override fun processIsolation(): WorkQueue {
+        TODO("Not yet implemented")
+    }
+
+    override fun processIsolation(action: Action<in ProcessWorkerSpec>?): WorkQueue {
+        TODO("Not yet implemented")
+    }
+
+    override fun await() {}
+}
+
+private class TestCheckNpmPackageAction(
+    project: Project,
+) : CheckNpmPackageAction() {
+    val params =
+        object : Params {
+            override val dependencyName: Property<String> = project.objects.property()
+
+            override val dependencyVersion: Property<String> =
+                project.objects.property()
+
+            override val outputDirectory: DirectoryProperty =
+                project.objects.directoryProperty()
+
+            override val networkService: Property<NetworkService> =
+                project.objects.property<NetworkService>().value(TestNetworkService(project))
+        }
+
+    override fun getParameters(): Params = params
+}
+
+private class TestNetworkService(
+    project: Project,
+) : NetworkService() {
+    val params =
+        object : Params {
+            override val baseUrl: Property<String> = project.objects.property<String>().value("https://cmgapps.com")
+            override val additionalHeaders: MapProperty<String, String> = project.objects.mapProperty()
+            override val engine: Property<HttpClientEngine> =
+                project.objects.property<HttpClientEngine>().value(
+                    MockEngine { request ->
+                        when {
+                            request.url.pathSegments[0] == TEST_DEP_NAME ->
+                                respond(
+                                    content =
+                                        ByteReadChannel(
+                                            Json.encodeToString(
+                                                NpmResponse(
+                                                    TEST_DEP_NAME,
+                                                    TEST_DEP_VERSION,
+                                                ),
+                                            ),
+                                        ),
+                                    status = HttpStatusCode.OK,
+                                    headers = headersOf(HttpHeaders.ContentType, "application/json"),
+                                )
+
+                            else ->
+                                respond(
+                                    "Not Found",
+                                    HttpStatusCode.NotFound,
+                                )
+                        }
+                    },
+                )
+        }
+
+    override fun getParameters(): Params = params
 }