+++ src/test/kotlin/com/cmgapps/gradle/reporter/OutputStreamTest.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2024. Christian Grach <christian.grach@cmgapps.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.cmgapps.gradle.reporter
+
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import java.io.ByteArrayOutputStream
+
+abstract class OutputStreamTest {
+    protected lateinit var outputStream: ByteArrayOutputStream
+
+    @BeforeEach
+    fun setUp() {
+        outputStream = ByteArrayOutputStream()
+    }
+
+    @AfterEach
+    fun tearDown() {
+        outputStream.close()
+    }
+}
+
+fun ByteArrayOutputStream.asString() = String(this.toByteArray())