+++ src/test/kotlin/com/cmgapps/gradle/reporter/TextReporterShould.kt
@@ -3,53 +3,32 @@ package com.cmgapps.gradle.reporter
 import com.cmgapps.gradle.model.Package
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.`is`
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.io.ByteArrayOutputStream
 
-class TextReporterShould {
-    private lateinit var outputStream: ByteArrayOutputStream
-
-    @BeforeEach
-    fun setUp() {
-        outputStream = ByteArrayOutputStream()
-    }
-
-    @AfterEach
-    fun tearDown() {
-        outputStream.close()
-    }
-
+class TextReporterShould : OutputStreamTest() {
     @Test
     fun `should print header`() {
         TextReporter(emptyList(), emptyList()).writePackages(outputStream)
 
-        assertThat(String(outputStream.toByteArray()), `is`("┌──────────────┐\n│ NPM Packages │\n└──────────────┘\n\n"))
+        assertThat(
+            outputStream.asString(),
+            `is`(
+                "┌──────────────┐\n" +
+                    "│ NPM Packages │\n" +
+                    "└──────────────┘\n\n",
+            ),
+        )
     }
 
     @Test
     fun `should print outdated and latest`() {
         TextReporter(
-            outdated =
-                listOf(
-                    Package(
-                        name = "outdated lib",
-                        currentVersion = "1.0.0",
-                        availableVersion = "2.0.0",
-                    ),
-                ),
-            listOf(
-                Package(
-                    name = "latest list",
-                    currentVersion = "1.0.0",
-                    availableVersion = "1.0.0",
-                ),
-            ),
+            outdated = outdated,
+            latest = latest,
         ).writePackages(outputStream)
 
         assertThat(
-            String(outputStream.toByteArray()),
+            outputStream.asString(),
             `is`(
                 "┌──────────────┐\n" +
                     "│ NPM Packages │\n" +
@@ -65,19 +44,12 @@ class TextReporterShould {
     @Test
     fun `should print outdated only`() {
         TextReporter(
-            outdated =
-                listOf(
-                    Package(
-                        name = "outdated lib",
-                        currentVersion = "1.0.0",
-                        availableVersion = "2.0.0",
-                    ),
-                ),
+            outdated = outdated,
             emptyList(),
         ).writePackages(outputStream)
 
         assertThat(
-            String(outputStream.toByteArray()),
+            outputStream.asString(),
             `is`(
                 "┌──────────────┐\n" +
                     "│ NPM Packages │\n" +